fix shadow crud breaking build when resolvers are not defined in a plugin
