+++ packages/strapi-plugin-graphql/services/utils.js
@@ -87,6 +87,85 @@ const amountLimiting = (params = {}) => {
 
 const nonRequired = type => type.replace('!', '');
 
+const actionExists = ({ resolver, resolverOf }) => {
+  if (isResolvablePath(resolverOf)) {
+    return true;
+  } else if (_.isFunction(resolver)) {
+    return true;
+  } else if (_.isString(resolver)) {
+    return _.isFunction(getActionFn(getActionDetails(resolver)));
+  } else {
+    throw new Error(
+      `Error building query. Expected \`resolver\` as string or a function, or \`resolverOf\` as a string. got ${{
+        resolver,
+        resolverOf,
+      }}`
+    );
+  }
+};
+
+const getAction = resolver => {
+  if (!_.isString(resolver)) {
+    throw new Error(`Error building query. Expected a string, got ${resolver}`);
+  }
+
+  const actionDetails = getActionDetails(resolver);
+  const actionFn = getActionFn(actionDetails);
+
+  if (!actionFn) {
+    throw new Error(
+      `[GraphQL] Cannot find action "${resolver}". Check your graphql configurations.`
+    );
+  }
+
+  return actionFn;
+};
+
+const getActionFn = details => {
+  const { controller, action, plugin, api } = details;
+
+  if (plugin) {
+    return _.get(strapi.plugins, [_.toLower(plugin), 'controllers', _.toLower(controller), action]);
+  }
+
+  return _.get(strapi.api, [_.toLower(api), 'controllers', _.toLower(controller), action]);
+};
+
+const getActionDetails = resolver => {
+  if (resolver.startsWith('plugins::')) {
+    const [, path] = resolver.split('::');
+    const [plugin, controller, action] = path.split('.');
+
+    return { plugin, controller, action };
+  }
+
+  if (resolver.startsWith('application::')) {
+    const [, path] = resolver.split('::');
+    const [api, controller, action] = path.split('.');
+
+    return { api, controller, action };
+  }
+
+  const args = resolver.split('.');
+
+  if (args.length === 3) {
+    const [api, controller, action] = args;
+    return { api, controller, action };
+  }
+
+  // if direct api access
+  if (args.length === 2) {
+    const [controller, action] = args;
+    return { api: controller, controller, action };
+  }
+
+  throw new Error(
+    `[GraphQL] Could not find action for resolver "${resolver}". Check your graphql configurations.`
+  );
+};
+
+const isResolvablePath = path => _.isString(path) && !_.isEmpty(path);
+
 module.exports = {
   diffResolvers,
   mergeSchemas,
@@ -95,4 +174,9 @@ module.exports = {
   convertToQuery,
   amountLimiting,
   nonRequired,
+  actionExists,
+  getAction,
+  getActionDetails,
+  getActionFn,
+  isResolvablePath,
 };
